Refactor app and login components with Redux hooks
