+++ ui/src/components/PrivateOutlet.jsx
@@ -1,0 +1,16 @@
+import React from 'react';
+import { useSelector } from 'react-redux';
+import { Navigate, Outlet, useLocation } from 'react-router-dom';
+
+function PrivateOutlet() {
+  const loggedIn = useSelector((state) => state.login.loggedIn);
+  const location = useLocation();
+
+  return loggedIn ? (
+    <Outlet />
+  ) : (
+    <Navigate to="/login" state={{ from: location }} replace />
+  );
+}
+
+export default PrivateOutlet;
