+++ ui/src/containers/LoginContainer.jsx
@@ -1,32 +1,16 @@
 import React from 'react';
-import { useLocation, Navigate } from 'react-router-dom';
-import { bindActionCreators } from 'redux';
-import { connect } from 'react-redux';
-import { logIn } from '../actions/login';
-import Login from '../components/Login/Login';
+import { Navigate } from 'react-router-dom';
+import { useSelector } from 'react-redux';
+import LoginForm from '../components/LoginForm/LoginForm';
 
-function LoginContainer(props) {
-  const location = useLocation();
+function LoginContainer() {
+  const loggedIn = useSelector((state) => state.login.loggedIn);
 
-  return props.data.loggedIn === false ? (
-    <Login {...props} />
-  ) : (
-    <Navigate to="/datacollection" state={{ from: location }} replace />
-  );
+  if (loggedIn) {
+    return <Navigate to="/datacollection" replace />;
+  }
+
+  return <LoginForm />;
 }
 
-function mapStateToProps(state) {
-  return {
-    showError: state.general.showErrorPanel,
-    errorMessage: state.general.errorMessage,
-    data: state.login,
-  };
-}
-
-function mapDispatchToProps(dispatch) {
-  return {
-    logIn: bindActionCreators(logIn, dispatch),
-  };
-}
-
-export default connect(mapStateToProps, mapDispatchToProps)(LoginContainer);
+export default LoginContainer;
